pull from base

+++ Commands/Lists/RemoveList.cs
@@ -1,20 +1,34 @@
-﻿using OfficeDevPnP.PowerShell.Commands.Base.PipeBinds;
+﻿using System.Management.Automation;
 using Microsoft.SharePoint.Client;
-using System.Management.Automation;
-using OfficeDevPnP.PowerShell.CmdletHelpAttributes;
+using SharePointPnP.PowerShell.CmdletHelpAttributes;
+using SharePointPnP.PowerShell.Commands.Base.PipeBinds;
 
-
-namespace OfficeDevPnP.PowerShell.Commands
+namespace SharePointPnP.PowerShell.Commands.Lists
 {
-    [Cmdlet(VerbsCommon.Remove, "SPOList", SupportsShouldProcess = true)]
+    [Cmdlet(VerbsCommon.Remove, "PnPList", SupportsShouldProcess = true)]
     [CmdletHelp("Deletes a list",
         Category = CmdletHelpCategory.Lists)]
-    public class RemoveList : SPOWebCmdlet
+    [CmdletExample(
+        Code = "PS:> Remove-PnPList -Identity Announcements",
+        SortOrder = 1,
+        Remarks = @"Removes the list named 'Announcements'. Asks for confirmation.")]
+    [CmdletExample(
+        Code = "PS:> Remove-PnPList -Identity Announcements -Force",
+        SortOrder = 2,
+        Remarks = @"Removes the list named 'Announcements' without asking for confirmation.")]
+    [CmdletExample(
+        Code = "PS:> Remove-PnPList -Title Announcements -Recycle",
+        SortOrder = 3,
+        Remarks = @"Removes the list named 'Announcements' and saves to the Recycle Bin")]
+    public class RemoveList : PnPWebCmdlet
     {
         [Parameter(Mandatory = true, ValueFromPipeline = true, Position = 0, HelpMessage = "The ID or Title of the list.")]
         public ListPipeBind Identity = new ListPipeBind();
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, HelpMessage = "Defines if the list should be moved to recycle bin or directly deleted.")]
+        public SwitchParameter Recycle;
+
+        [Parameter(Mandatory = false, HelpMessage = "Specifying the Force parameter will skip the confirmation question.")]
         public SwitchParameter Force;
         protected override void ExecuteCmdlet()
         {
@@ -25,7 +39,14 @@ namespace OfficeDevPnP.PowerShell.Commands
                 {
                     if (Force || ShouldContinue(Properties.Resources.RemoveList, Properties.Resources.Confirm))
                     {
-                        list.DeleteObject();
+                        if (Recycle)
+                        {
+                            list.Recycle();
+                        }
+                        else
+                        {
+                            list.DeleteObject();
+                        }
                         ClientContext.ExecuteQueryRetry();
                     }
                 }
